lab03/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiokafka==0.12.0
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
click==8.1.8
colorlog==6.9.0
confluent-kafka==2.11.0
croniter==2.0.7
faust-streaming==0.11.3
frozenlist==1.7.0
idna==3.10
intervaltree==3.1.0
mode-streaming==0.4.1
multidict==6.6.3
mypy_extensions==1.1.0
opentracing==2.4.0
packaging==25.0
propcache==0.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
rocksdict==0.3.27
six==1.17.0
sortedcontainers==2.4.0
terminaltables==3.1.10
typing_extensions==4.14.1
venusian==3.1.0
yarl==1.20.1
